Extension: Result from second activity

+++ app/src/main/java/com/assignment/trial5_core2/member.kt
@@ -7,6 +7,6 @@ import kotlinx.android.parcel.Parcelize
 @Parcelize
 data class member( val name: String, val stagename: String,
                    val nickname: String, val age: Int, val Role: String, val knownFor: String,
-                   var rating : Int, val img : String, var KnowHim: Boolean): Parcelable
+                   var rating : Float, val img : String, var KnowHim: Boolean): Parcelable
 {
 }
